gremlinFlip Minigame
Inspired by Voltorb Flip, a minigame in Pokemon HeartGold and SoulSilver, Gremlin Flip tasks the player with flipping cards to gain gold and avoiding Gremlin Nobs.

The first tile flipped gives the player that much gold on the first card, and each subsequent card that is greater than 1 multiplies the total by the number on the card.
For subsequent tiles that are flipped that have the number 1, 1 is added to the player's gold.

Flipping a Gremlin Nob causes the play to spill all gold and lose the minigame.

The player presses left-click to flip tiles, and can press right-click to mark a tile.